Add compatibility layer to separate v1 and v2.
