+++ django_pydantic_field/compat/__init__.py
@@ -1,0 +1,5 @@
+from .pydantic import PYDANTIC_V1 as PYDANTIC_V1
+from .pydantic import PYDANTIC_V2 as PYDANTIC_V2
+from .pydantic import PYDANTIC_VERSION as PYDANTIC_VERSION
+from .django import GenericContainer as GenericContainer
+from .django import MigrationWriter as MigrationWriter
