+++ django_pydantic_field/v1/__init__.py
@@ -1,0 +1,1 @@
+from .fields import *
